chore: Moved legacy mapper to config project
