+++ ra2ce/analysis/analysis_config_data/damages_config_data.py
@@ -31,8 +31,8 @@ from ra2ce.analysis.analysis_config_data.enums.event_type_enum import EventTypeE
 from ra2ce.analysis.analysis_config_data.enums.risk_calculation_mode_enum import (
     RiskCalculationModeEnum,
 )
-from ra2ce.analysis.analysis_config_data.legacy_mappers import with_legacy_mappers
 from ra2ce.common.validation.validation_report import ValidationReport
+from ra2ce.configuration.legacy_mappers import with_legacy_mappers
 
 
 @with_legacy_mappers
