+++ ra2ce/analysis/analysis_config_data/single_link_redundancy_config_data.py
@@ -25,7 +25,7 @@ from ra2ce.analysis.analysis_config_data.analysis_config_data_protocol import (
     AnalysisConfigDataProtocol,
 )
 from ra2ce.analysis.analysis_config_data.enums.weighing_enum import WeighingEnum
-from ra2ce.analysis.analysis_config_data.legacy_mappers import with_legacy_mappers
+from ra2ce.configuration.legacy_mappers import with_legacy_mappers
 
 
 @with_legacy_mappers
